verify approval and recovery races through their real entry points
